fix: 解决合并冲突并修复启动崩溃问题

+++ app/src/main/java/com/csbaby/kefu/infrastructure/reply/ReplyOrchestrator.kt
@@ -345,7 +345,7 @@ class ReplyOrchestrator @Inject constructor(
             accepted = !modified,
             modified = modified,
             rejected = false,
-            confidence = confidence
+            confidence = confidence.toDouble()
         )
     }
